requirements.txt
streamlit
google-generativeai>=0.8.4
chromadb
pillow